+++ bhlindex.go
@@ -3,7 +3,9 @@ package bhlindex
 import (
 	"database/sql"
 	"fmt"
+	"log"
 	"os"
+	"strconv"
 	"strings"
 
 	uuid "github.com/satori/go.uuid"
@@ -11,10 +13,11 @@ import (
 
 // Env is a collection of environment variables.
 type Env struct {
-	DbHost string
-	DbUser string
-	Db     string
-	BHLDir string
+	DbHost      string
+	DbUser      string
+	Db          string
+	BHLDir      string
+	PrefSources []int
 }
 
 // Check handles error checking, and panicks if error is not nil.
@@ -27,8 +30,8 @@ func Check(err error) {
 // EnvVars imports all environment variables relevant for the data conversion.
 func EnvVars() Env {
 	emptyEnvs := make([]string, 0, 4)
-	envVars := [4]string{"POSTGRES_HOST", "POSTGRES_USER", "POSTGRES_DB",
-		"BHL_DIR"}
+	envVars := [5]string{"POSTGRES_HOST", "POSTGRES_USER", "POSTGRES_DB",
+		"BHL_DIR", "PREF_SOURCES"}
 	for i, v := range envVars {
 		val, ok := os.LookupEnv(v)
 		if ok {
@@ -41,8 +44,20 @@ func EnvVars() Env {
 		envs := strings.Join(emptyEnvs, ", ")
 		panic(fmt.Errorf("Environment variables %s are not defined", envs))
 	}
+	var sources []int
+	for _, v := range strings.Split(envVars[4], ",") {
+		v = strings.Trim(v, " ")
+		source, err := strconv.Atoi(v)
+		if err != nil {
+			log.Println(err)
+			log.Fatal(`
+PREF_SOURCES env variable should be a comma-separated list of ints:
+Example: "1,2,3,4"`)
+		}
+		sources = append(sources, source)
+	}
 	return Env{DbHost: envVars[0], DbUser: envVars[1], Db: envVars[2],
-		BHLDir: envVars[3]}
+		BHLDir: envVars[3], PrefSources: sources}
 }
 
 // UUID4 returns random (version 4) UUID as a string.